Little fix
We know the bootstrap is loaded by default in J3 but sometimes, the
scripts loading order is messed up and this little fix make everything
working correctly (so firstly is loaded bootstrap and THEN the JDOM
extra JS libraries).
the ideal solution would be a better joomla scripts loading system.
Until now nobody seems to care about this.